+++ test/base/test_objective.py
@@ -3,11 +3,12 @@ This is for testing the pypesto.Objective.
 """
 
 import numpy as np
-import scipy.optimize as so
 import sympy as sp
 import numbers
 import pytest
 import pypesto
+
+from ..util import rosen_for_sensi, poly_for_sensi
 
 
 @pytest.fixture(params=[True, False])
@@ -125,95 +126,6 @@ def _test_sensis(struct):
             obj(x, (0, 1, 2))
 
 
-def obj_for_sensi(fun, grad, hess, max_sensi_order, integrated, x):
-    """
-    Create a pypesto.Objective able to compute up to the speficied
-    max_sensi_order. Returns a dict containing the objective obj as well
-    as max_sensi_order and fval, grad, hess for the passed x.
-
-    Parameters
-    ----------
-
-    fun, grad, hess: callable
-        Functions computing the fval, grad, hess.
-    max_sensi_order: int
-        Maximum sensitivity order the pypesto.Objective should be capable of.
-    integrated: bool
-        True if fun, grad, hess should be integrated into one function, or
-        passed to pypesto.Objective separately (both is possible)
-    x: np.array
-        Value at which to evaluate the function to obtain true values.
-
-    Returns
-    -------
-
-    ret: dict
-        With fields obj, max_sensi_order, x, fval, grad, hess.
-    """
-    if integrated:
-        if max_sensi_order == 2:
-            def arg_fun(x):
-                return fun(x), grad(x), hess(x)
-            arg_grad = arg_hess = True
-        elif max_sensi_order == 1:
-            def arg_fun(x):
-                return fun(x), grad(x)
-            arg_grad = True
-            arg_hess = False
-        else:
-            def arg_fun(x):
-                return fun(x)
-            arg_grad = arg_hess = False
-    else:  # integrated
-        if max_sensi_order >= 2:
-            arg_hess = hess
-        else:
-            arg_hess = None
-        if max_sensi_order >= 1:
-            arg_grad = grad
-        else:
-            arg_grad = None
-        arg_fun = fun
-    obj = pypesto.Objective(fun=arg_fun, grad=arg_grad, hess=arg_hess)
-    return {'obj': obj,
-            'max_sensi_order': max_sensi_order,
-            'x': x,
-            'fval': fun(x),
-            'grad': grad(x),
-            'hess': hess(x)}
-
-
-def rosen_for_sensi(max_sensi_order, integrated=False, x=None):
-    """
-    Rosenbrock function from scipy.optimize.
-    """
-    if x is None:
-        x = [0, 1]
-
-    return obj_for_sensi(so.rosen,
-                         so.rosen_der,
-                         so.rosen_hess,
-                         max_sensi_order, integrated, x)
-
-
-def poly_for_sensi(max_sensi_order, integrated=False, x=0.):
-    """
-    1-dim polynomial for testing in 1d.
-    """
-
-    def fun(x):
-        return (x - 2)**2 + 1
-
-    def grad(x):
-        return 2 * (x - 2)
-
-    def hess(_):
-        return 2
-
-    return obj_for_sensi(fun, grad, hess,
-                         max_sensi_order, integrated, x)
-
-
 def test_finite_difference_checks():
     """
     Test the finite difference gradient check methods by expected relative
@@ -231,15 +143,15 @@ def test_finite_difference_checks():
 
     objective = pypesto.Objective(fun=fun, grad=grad)
 
-    def rel_err(eps):
+    def rel_err(eps_):
         """Expected relative error."""
-        central_difference = (fun(theta + eps) - fun(theta - eps))/(2*eps)
-        return abs(
-                (grad(theta) - central_difference)/(central_difference + eps))
+        central_difference = (fun(theta + eps_) - fun(theta - eps_))/(2*eps_)
+        return abs((grad(theta) - central_difference) /
+                   (central_difference + eps_))
 
     # Test the single step size `check_grad` method.
     eps = 1e-5
-    result_single_eps = objective.check_grad([theta], eps=eps)
+    result_single_eps = objective.check_grad(np.array([theta]), eps=eps)
     assert result_single_eps['rel_err'].squeeze() == rel_err(eps)
 
     # Test the multiple step size `check_grad_multi_eps` method.
