+++ test/petab/test_amici_objective.py
@@ -10,7 +10,7 @@ import pypesto.optimize
 import pypesto.objective.constants
 import pytest
 import numpy as np
-from test.petab_util import folder_base
+from .petab_util import folder_base
 
 ATOL = 1e-1
 RTOL = 1e-0
