+++ test/visualize/__init__.py
@@ -1,3 +1,4 @@
 """Visualization tests."""
 
-from .test_visualize import close_fig
+from .test_visualize import (
+    close_fig, create_problem, create_optimization_result)
